+++ DataAccess/Operations/EmployeeShiftOperations.cs
@@ -20,7 +20,8 @@ namespace DataAccess.Operations
         }
         public EmployeesShifts DeleteEmployeeShift(int id)
         {
-            var employeeShift = GetByPk(id);
+            EmployeesShifts employeeShift = GetByPk(id);
+            if (employeeShift == null) { return null; }
             db.Remove(employeeShift);
             db.SaveChanges();
             return employeeShift;
